fix : modify return type of update and delete method

+++ todo/src/main/java/jungle/study/todo/api/service/TodoService.java
@@ -10,8 +10,8 @@ import java.util.List;
 public interface TodoService {
 
     Todo registerTodo(TodoDto todoDto);
-    int updateTodo(Long todoId, TodoDto todoDto);
-    int deleteTodo(Long todoId);
+    Long updateTodo(Long todoId, TodoDto todoDto);
+    Long deleteTodo(Long todoId);
     Todo findTodoById(Long todoId);
     List<Todo> findAll();
 }
